Added styles to the footer
Changed the cualquiera .introducction
Created the class main-nav (commented)
Created the section footer where I included: 
the main-fotter and a text-align, bottom-footer-left and a text align to left, a bottom-footer-rigth and a text align to rigth.
Added the line above the footer targin footer and adding a border-top.
Did SOME TASTES with media queries.